refactor(holdings): collapse duplicated backtest price-load orchestration into BacktestPriceLoader.RunBacktest
